tools/book-processor/requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
nltk==3.9
sqlite3
flask==3.0.0
whoosh==2.7.4
textract==1.6.5
python-docx==1.1.0
beautifulsoup4==4.12.2
networkx==3.2.1
matplotlib==3.8.2
plotly==5.17.0
pandas==2.1.4